refactor(export): move the STEP export vocabulary out of step-exporter (#2475)
step-exporter.ts is 1169 lines against a ~400 rule. This is the first of
several cuts, and deliberately the one that cannot change behaviour:
StepExportOptions, StepExportProgress, StepExportResult,
SourceLineMutations, ExportPass and the relationshipWithheldWarning
message builder move to step-export-types.ts. Declarations only. No guard,
no predicate and no control flow moved.

1169 -> 945.

Nothing else moved with them, on purpose. step-exporter.ts re-exports all
five types, so the seven sibling modules that import ExportPass or
SourceLineMutations from it are untouched, and packages/export/src/index.ts
line 16 -- which is the package's public surface for
StepExportOptions/Result/Progress -- is byte-identical. A consumer cannot
tell this happened.

Why this cut first: the issue's binding constraint is that the guards are
pinned by mutation tests mapped to specific call sites, and a split that
detaches a guard from its test leaves the suite green for the wrong reason.
Type declarations have no such coupling, so this establishes the new file
and the re-export seam with no risk, before anything load-bearing moves.

One non-obvious import: ExportPass types isReadableSourceRef as
ReturnType<typeof createSourceRefReader> rather than restating the shape,
so the new module needs createSourceRefReader as a value-position import
even though nothing here calls it.

The allowlist row drops 1169 -> 945 with the digest re-pinned. The
regenerator also wanted to lower two unrelated rows that already carried
slack on main (headless-backend.ts 778->752, backend-query.ts 509->485);
those are restored to their current values and the digest computed from the
restored file. Lowering a budget is never a weakening, but those files are
not this PR's to tighten -- a concurrent PR growing either one inside its
existing budget should not go red because of a refactor elsewhere.

Verified against the pre-change baseline, same command both times:
  pnpm exec turbo run test --filter=@ifc-lite/export --force
  before 63 files / 881 tests   after 63 files / 881 tests
Dependents: @ifc-lite/mutations 218/218, @ifc-lite/sandbox 200/200.
check-module-size exit 0, 0 new over 400; step-export-types.ts is 274
lines and needs no row.

Still to come for #2475, with sizes measured rather than guessed: the
ExportPass literal (250 lines, contiguous), the predicate block (263, and
it contains an early return out of export() that has to become a returned
value), and the five context builders. The file lands near 310 after all
three -- so the ~400 target is reachable, but not in one move.
